requirements.txt
Django
django-allauth
drf-yasg
djangorestframework
dj-rest-auth
Pillow


